Simplify filtering for suitable labels.
The "designation IS NULL" predicate was a stop-gap solution to avoid
catching PE symbol elements with empty label, which had to be marked
manually with a non-emtpy article number field.

The SQL string splitting functions already require the colon, so
simply filter out any terminals that don't meet this requirement.